Module monster::incubation

This module includes new utilities which are being matured and tested for permanent inclusion. Items in this module should be considered unstable and thus may change or get removed completely at any time.

Reexports

pub use self::swap_tuple::SwapTupleExt;

pub use self::map_ref_mut::{MapRefMutExt, map_ref_mut};

pub use self::zip_ends::{ZipEnds, ZipEndsIterExt};

pub use self::fmt::{format, write_format};

pub use self::some_if::some_if;

pub use self::tuple_indexing::TupleIndexingExt;

pub use self::tuple_iter::TupleIterExt;

pub use self::ref_eq::RefEq;

pub use self::global_init_cleanup::{InitCleanup, GlobalInitHandle, InitHandle};

pub use self::with_lock::WithLock;

pub use self::find_and_take::FindAndTake;

pub use self::slice_drop_first::SliceDropFirst;

pub use self::slice_drop_last::SliceDropLast;

pub use self::slice_take_while::SliceTakeWhile;

pub use self::split_int::{SplitInt, split_mut_u16};

pub use self::reverse_bitorder::{ReverseBitorder, reverse_bitorder};

pub use self::owning_ref_mut::OwningRefMut;

Modules

anon

find_and_take

fmt

global_init_cleanup

map_ref_mut

mutate

owning_ref_mut

ref_eq

reverse_bitorder

slice_drop_first

slice_drop_last

slice_take_while

some_if

split_int

swap_tuple

tuple_indexing

tuple_iter

with_lock

zip_ends